+++ front/components/actions/mcp/ConnectMCPServerDialog.tsx
@@ -21,7 +21,7 @@ import {
   isRemoteMCPServerType,
 } from "@app/lib/actions/mcp_helper";
 import { DEFAULT_MCP_SERVER_ICON } from "@app/lib/actions/mcp_icons";
-import type { AuthorizationInfo } from "@app/lib/actions/mcp_metadata";
+import type { AuthorizationInfo } from "@app/lib/actions/mcp_metadata_extraction";
 import type { MCPServerViewType } from "@app/lib/api/mcp";
 import {
   useCreateMCPServerConnection,
