+++ front/lib/actions/mcp_metadata.ts
@@ -5,16 +5,10 @@ import type { StreamableHTTPClientTransportOptions } from "@modelcontextprotocol
 import { StreamableHTTPClientTransport } from "@modelcontextprotocol/sdk/client/streamableHttp.js";
 import type { AuthInfo } from "@modelcontextprotocol/sdk/server/auth/types.js";
 import type { OAuthTokens } from "@modelcontextprotocol/sdk/shared/auth.js";
-import type { Implementation, Tool } from "@modelcontextprotocol/sdk/types.js";
+import type { Tool } from "@modelcontextprotocol/sdk/types.js";
 import type { JSONSchema7 as JSONSchema } from "json-schema";
 import { ProxyAgent } from "undici";
 
-import { isInternalAllowedIcon } from "@app/components/resources/resources_icons";
-import {
-  DEFAULT_MCP_ACTION_DESCRIPTION,
-  DEFAULT_MCP_ACTION_NAME,
-  DEFAULT_MCP_ACTION_VERSION,
-} from "@app/lib/actions/constants";
 import {
   getConnectionForMCPServer,
   MCPServerPersonalAuthenticationRequiredError,
@@ -25,65 +19,28 @@ import {
   doesInternalMCPServerRequireBearerToken,
   getServerTypeAndIdFromSId,
 } from "@app/lib/actions/mcp_helper";
-import { DEFAULT_MCP_SERVER_ICON } from "@app/lib/actions/mcp_icons";
 import { connectToInternalMCPServer } from "@app/lib/actions/mcp_internal_actions";
 import { InMemoryWithAuthTransport } from "@app/lib/actions/mcp_internal_actions/in_memory_with_auth_transport";
+import { extractMetadataFromServerVersion } from "@app/lib/actions/mcp_metadata_extraction";
 import { MCPOAuthRequiredError } from "@app/lib/actions/mcp_oauth_error";
 import { MCPOAuthProvider } from "@app/lib/actions/mcp_oauth_provider";
 import type { AgentLoopContextType } from "@app/lib/actions/types";
 import { ClientSideRedisMCPTransport } from "@app/lib/api/actions/mcp_client_side";
-import type {
-  InternalMCPServerDefinitionType,
-  MCPServerDefinitionType,
-  MCPServerType,
-  MCPToolType,
-} from "@app/lib/api/mcp";
+import type { MCPServerType, MCPToolType } from "@app/lib/api/mcp";
 import type { Authenticator } from "@app/lib/auth";
 import { getUntrustedEgressAgent } from "@app/lib/egress/server";
 import { isWorkspaceUsingStaticIP } from "@app/lib/misc";
 import { InternalMCPServerCredentialModel } from "@app/lib/models/agent/actions/internal_mcp_server_credentials";
 import { RemoteMCPServerResource } from "@app/lib/resources/remote_mcp_servers_resource";
 import logger from "@app/logger/logger";
-import type { MCPOAuthUseCase, OAuthProvider, Result } from "@app/types";
+import type { MCPOAuthUseCase, Result } from "@app/types";
 import {
   assertNever,
   EnvironmentConfig,
   Err,
-  isOAuthProvider,
   normalizeError,
   Ok,
 } from "@app/types";
-
-export type AuthorizationInfo = {
-  provider: OAuthProvider;
-  supported_use_cases: MCPOAuthUseCase[];
-  scope?: string;
-};
-
-export function isAuthorizationInfo(a: unknown): a is AuthorizationInfo {
-  return (
-    typeof a === "object" &&
-    a !== null &&
-    "provider" in a &&
-    isOAuthProvider(a.provider) &&
-    "supported_use_cases" in a
-  );
-}
-
-export function isInternalMCPServerDefinition(
-  server: Implementation
-): server is InternalMCPServerDefinitionType {
-  return (
-    "authorization" in server &&
-    (isAuthorizationInfo(server.authorization) ||
-      server.authorization === null) &&
-    "description" in server &&
-    typeof server.description === "string" &&
-    "icon" in server &&
-    typeof server.icon === "string" &&
-    isInternalAllowedIcon(server.icon)
-  );
-}
 
 interface ConnectViaMCPServerId {
   type: "mcpServerId";
@@ -505,40 +462,6 @@ async function connectToRemoteMCPServer(
       throw error;
     }
   }
-}
-
-export function extractMetadataFromServerVersion(
-  r: Implementation | undefined
-): MCPServerDefinitionType {
-  if (r) {
-    return {
-      name: r.name ?? DEFAULT_MCP_ACTION_NAME,
-      version: r.version ?? DEFAULT_MCP_ACTION_VERSION,
-      authorization: isInternalMCPServerDefinition(r) ? r.authorization : null,
-      description: isInternalMCPServerDefinition(r)
-        ? r.description
-        : DEFAULT_MCP_ACTION_DESCRIPTION,
-      icon: isInternalMCPServerDefinition(r) ? r.icon : DEFAULT_MCP_SERVER_ICON,
-      documentationUrl: isInternalMCPServerDefinition(r)
-        ? r.documentationUrl
-        : null,
-      developerSecretSelection: isInternalMCPServerDefinition(r)
-        ? r.developerSecretSelection
-        : undefined,
-      developerSecretSelectionDescription: isInternalMCPServerDefinition(r)
-        ? r.developerSecretSelectionDescription
-        : undefined,
-    };
-  }
-
-  return {
-    name: DEFAULT_MCP_ACTION_NAME,
-    version: DEFAULT_MCP_ACTION_VERSION,
-    description: DEFAULT_MCP_ACTION_DESCRIPTION,
-    icon: DEFAULT_MCP_SERVER_ICON,
-    authorization: null,
-    documentationUrl: null,
-  };
 }
 
 export function extractMetadataFromTools(tools: Tool[]): MCPToolType[] {
